ThemeContext: add global theme feature
This commit adds a global ThemeContext that provides a color theme,
which is selected from theme.json. From theme.json, either schemes.light
or schemes.dark is provided as the colorTheme passed to the children,
depending on the device's current color theme. A hook is used to detect
color theme changes and will update the UI accordingly.

All components have been updated to use this color theme, other than the
hardcoded color scales for SPE and ORMA. New nodes are planned for the
colorTheme to accomodate these. These nodes will likely not be added
directly to theme.json to make the file easy to update. The file is
generated at
https://material-foundation.github.io/material-theme-builder/.