github/workflows: set up automatic Firebase deploy
This commit adds two GitHub workflow files. One triggers on each pull
request, building and deploying the contents to Firebase and generating
a preview link that will expire in 7 days. The other will trigger when
commits are added to the main branch (merges) and will build and deploy
the web app to production (tools.ca-sar.org).

Can't test this very well because it is a workflow file. Going to push
and merge to test and may have to follow up later.