Branding changes and iOS fix
This commit updates the branding of the app and fixes functionality on
iOS devices.

At the request of Mike, the app has been updated to use the CALSAR logo
directly. The CALSAR wordmark is currently not approved, but is being
used in the interm. For clarity, the app is being renamed from opsRisk
to Risk. The NPS Risk app isn't available anymore, so there shouldn't be
much confusion.

A workaround has been added for iOS, fixes the app freezing
when the ORMA page is opened. Still not super sure about the cause/a
proper fix as devtools isn't opening on my computer.

@react-native-community/slider has been fixed upstream, so package
versions have been updated.