add README and dynamic favicon
Add README. The newly added favicon checks the system theme and uses
a light icon if the system uses a dark theme to enhance visibility on
web. It won't work properly if the system theme is light but the browser
tab bar is dark, and vice-versa, but it should be a bit better for many
folks.