ops: reorganize and simplify header
Changed the header when a file is open to be more explicit about printer
state. Also showing the file sync/save state in anticipation of having
server sync features.

This commit also removes the file info button. Instead, when the file
name is clicked, the tab opens. This cleans up the UI a little but still
makes logical sense, in my opinion.