ops: switch to RxDB
This commit changes from saving files from asyncstorage to RxDB backed
by dexie. Because we are using dexie, native builds are now broken (not
that I've tried building for Android/iOS yet), but there are no
expectations of compiling these for now.
There are likely bugs that need to be fixed.